refactor: split prototype to a file
