+++ js/app.js
@@ -1,3 +1,5 @@
+import "./prototype.js" 
+
 var pairing = []; // an array to pair clicked cards
 var paired = []; // an array to store paired cards
 var count = 0; // count move(click two different cards)
@@ -175,27 +177,3 @@ function rate() {
   $('.fa-star').eq(stars).removeClass('fa').addClass('far')
 }
 
-// Shuffle function from https://stackoverflow.com/a/6274381/9984029
-Array.prototype.shuffle = function () {
-  for (let i = this.length - 1; i > 0; i--) {
-    const j = Math.floor(Math.random() * (i + 1));
-    [this[i], this[j]] = [this[j], this[i]];
-  }
-
-  return this;
-};
-
-/**
- * Check whether current card is in cards array
- * it will return false when cards array is empty 
- * for loop will check each card in cards array
- *  array.includes() doesn't work here.
- * @param {jQuery Object} card 
- */
-Array.prototype.has = function (card) {
-  for (const pairedCard of this) {
-    if (isMatch(pairedCard, card)) return true;
-  }
-
-  return false;
-};
